+++ src/Resources/contao/config/config.php
@@ -2,4 +2,5 @@
 
 $GLOBALS["BE_MOD"]["content"]["schiedsrichterverteiler"] = array(
 	"tables"      => array('tl_schiedsrichterverteiler'),
+	"setDefault"  => array("Schachbulle\ContaoSchiedsrichterverteilerBundle\Classes\Standardverteiler","setDefault"),
 );