+++ src/Resources/contao/languages/de/default.php
@@ -1,0 +1,6 @@
+<?php
+
+$GLOBALS['TL_LANG']['tl_schiedsrichterverteiler']['setDefaultHeadline'] = array('Standardverteiler festlegen', 'Einen Standardverteiler festlegen');
+$GLOBALS['TL_LANG']['tl_schiedsrichterverteiler']['setDefaultSelectheadline'] = array('Verteiler auswählen', 'Bitte legen Sie einen Verteiler als Standardverteiler fest.');
+
+$GLOBALS['TL_LANG']['tl_schiedsrichterverteiler']['setDefaultSubmit'] = array('Verteiler zum Standard machen', 'Die Adressen im Schiedsrichter-Newsletter werden automatisch auf den eingestellten Verteiler aktualisiert.');