+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Schiedsrichter-Verteiler Changelog
 
+## Version 0.0.3 (2021-04-13)
+
+* Add: Funktionalität für Festlegung Standardverteiler
+* Add: Abhängigkeit schachbulle/contao-schiedsrichter-bundle ab Version 1.1
+* Add: Synchronisierung des Standardverteilers mit der Schiedsrichter-Datenbank
+
 ## Version 0.0.2 (2021-04-09)
 
 * Bundle ausgebaut mit den Grundfunktionen